@@ -4,7 +4,7 @@
   <parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		    <version>2.4.0</version>
+		    <version>3.0.6</version>
   </parent>
   
   <groupId>com.jdsalasca</groupId>
@@ -12,9 +12,9 @@
   <version>0.0.2</version>
    <url>http://maven.apache.org</url>
   <name>utils</name>
-  <description>utils for web development</description>
+  <description>utils for web development for spring boot 3 </description>
       <properties>
-		  	<java.version>11</java.version>
+		  	<java.version>17</java.version>
         <maven.compiler.target>1.8</maven.compiler.target>
         <maven.compiler.source>1.8</maven.compiler.source>
     </properties>
@@ -86,6 +86,7 @@
 	<distributionManagement>
    <repository>
      <id>github</id>
+     
      <name>GitHub jdsalasca Apache Maven Packages</name>
      <url>https://maven.pkg.github.com/jdsalasca/utils-web</url>
    </repository>
